+++ clang/test/Parser/cxx0x-attributes.cpp
@@ -71,11 +71,24 @@ namespace test_misplacement {
 [[]] union union_attr2; //expected-error{{misplaced attributes}}
 [[]] enum  E2 { }; //expected-error{{misplaced attributes}}
 }
-struct S1 { __attribute__((deprecated)) alignas(16) int x; }; // expected-none
-class C1 { __attribute__((deprecated)) alignas(16) int x; }; // expected-none
+
+__attribute__(()) [[]] alignas(int) int xx; // expected-none
+__attribute__(()) alignas(int) [[]] int yy; // expected-none
+[[]] __attribute__(()) alignas(int) int zz; // expected-none
+alignas(int) [[]] __attribute__(()) int aa; // expected-none
+
+class C1 {
+  __attribute__(()) [[]] alignas(int) int x; // expected-none
+  __attribute__(()) alignas(int) [[]] int y; // expected-none
+  [[]] __attribute__(()) alignas(int) int z; // expected-none
+  alignas(int) [[]] __attribute__(()) int a; // expected-none
+};
 
 void fn_with_decl() {
-  __attribute__((deprecated)) alignas(16) int x; // expected-none
+  __attribute__(()) alignas(int) int x; // expected-none
+  __attribute__(()) alignas(int) [[]] int y; // expected-none
+  [[]] __attribute__(()) alignas(int) int z; // expected-none
+  alignas(int) [[]] __attribute__(()) int a; // expected-none
 }
 
 // Checks attributes placed at wrong syntactic locations of class specifiers.